Add `npm run dev` and parcel for frontend skin development
Skin development is a lot easier if you can get out of PHP.
The skin is reduced to a mustache template, LESS template and JS template
and users can stay out of MediaWiki internals for the time being.

Note: This may be a little misleading as while imports in src will
work in parcel they won't work in MediaWiki.

Change-Id: Iaf59c8ada0a1be10d7f72bc0b2c6c80f265bfa16